+++ lib/plex_utils/plex_pair.dart
@@ -1,0 +1,8 @@
+class Pair<T1, T2> {
+  final T1 first;
+  final T2 second;
+
+  Pair._(this.first, this.second);
+
+  Pair.create(this.first, this.second);
+}